Rename module to more generic schema_exporter
